data/airflow-wordcloud/requirements.txt
alembic==1.12.0
anyio==4.0.0
apache-airflow==2.1.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.13.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.10.2
cattrs==1.10.0
certifi==2023.7.22
cffi==1.16.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.6
colorlog==6.7.0
croniter==1.0.15
cryptography==41.0.4
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.9.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
numpy==1.24.4
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==23.2
pandas==1.5.3
pendulum==2.1.2
pkg_resources==0.0.0
prison==0.2.1
psutil==5.9.5
py4j==0.10.9.2
pyarrow==13.0.0
pycparser==2.21
Pygments==2.16.1
PyJWT==1.7.1
PyKomoran==0.1.6.post1
pymongo==4.5.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rich==13.6.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==6.2.0
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.8.0
unicodecsv==0.14.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.17.0
